Working refactor. Separated out functions into bca.rs, material.rs, particle.rs, tests.rs, ...
