+++ src/tests.rs
@@ -1,0 +1,67 @@
+#[cfg(test)]
+use super::*;
+
+#[test]
+fn test_rotate_particle() {
+    let mass = 1.;
+    let Z = 1.;
+    let E = 1.;
+    let Ec = 1.;
+    let Es = 1.;
+    let x = 0.;
+    let y = 0.;
+    let z = 0.;
+    let cosx = (PI/4.).cos();
+    let cosy = (PI/4.).sin();
+    let cosz = 0.;
+    let psi = -PI/4.;
+    let phi = 0.;
+
+    let mut particle = particle::Particle::new(mass, Z, E, Ec, Es, x, y, z, cosx, cosy, cosz, false, false);
+
+    //Check that rotation in 2D works
+    rotate_particle(&mut particle, psi, phi);
+    assert!(approx_eq!(f64, particle.dir.x, 0., epsilon = 1e-9), "particle.dir.x: {} Should be ~0.", particle.dir.x);
+    assert!(approx_eq!(f64, particle.dir.y, 1., epsilon = 1e-9), "particle.dir.y: {} Should be ~1.", particle.dir.y);
+
+    //Check that rotating back by negative psi returns to the previous values
+    rotate_particle(&mut particle, -psi, phi);
+    assert!(approx_eq!(f64, particle.dir.x, cosx, epsilon = 1e-9), "particle.dir.x: {} Should be ~{}", particle.dir.x, cosx);
+    assert!(approx_eq!(f64, particle.dir.y, cosy, epsilon = 1e-9), "particle.dir.y: {} Should be ~{}", particle.dir.y, cosy);
+
+    //Check that azimuthal rotation by 180 degrees works correctly
+    let phi = PI;
+    rotate_particle(&mut particle, psi, phi);
+    assert!(approx_eq!(f64, particle.dir.x, 1., epsilon = 1e-9), "particle.dir.x: {} Should be ~1.", particle.dir.x);
+    assert!(approx_eq!(f64, particle.dir.y, 0., epsilon = 1e-9), "particle.dir.y: {} Should be ~0.", particle.dir.y);
+
+    //Check that particle direction vector remains normalized following rotations
+    assert!(approx_eq!(f64, particle.dir.x.powf(2.) + particle.dir.y.powf(2.) + particle.dir.z.powf(2.), 1.), "Particle direction not normalized.");
+
+}
+
+#[test]
+fn test_particle_advance() {
+    let mass = 1.;
+    let Z = 1.;
+    let E = 1.;
+    let Ec = 1.;
+    let Es = 1.;
+    let x = 0.;
+    let y = 0.;
+    let z = 0.;
+    let cosx = (PI/4.).cos();
+    let cosy = (PI/4.).sin();
+    let cosz = 0.;
+    let mfp = 1.;
+    let asymptotic_deflection = 0.5;
+
+    let mut particle = particle::Particle::new(mass, Z, E, Ec, Es, x, y, z, cosx, cosy, cosz, false, false);
+
+    let distance_traveled = particle_advance(&mut particle, mfp, asymptotic_deflection);
+
+    assert_eq!(particle.pos.x, (1. - 0.5)*cosx);
+    assert_eq!(particle.pos.y, (1. - 0.5)*cosy);
+    assert_eq!(particle.pos.z, 0.);
+    assert_eq!(distance_traveled, mfp - asymptotic_deflection);
+}
